+++ lib/Controller/LoginController.php
@@ -8,6 +8,7 @@ use OC\Authentication\Token\IProvider;
 use OC\User\Session as OC_UserSession;
 use OCA\HitobitoLogin\AppInfo\Application;
 use OCA\HitobitoLogin\Service\ProvisioningService;
+use OCA\HitobitoLogin\Service\SettingsService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\Attribute\FrontpageRoute;
@@ -48,6 +49,7 @@ class LoginController extends Controller {
 		private IUserManager $userManager,
 		private ProvisioningService $provisioningService,
 		private IURLGenerator $urlGenerator,
+		private SettingsService $settingsService,
 	) {
 		parent::__construct(Application::APP_ID, $request);
 	}
@@ -191,11 +193,22 @@ class LoginController extends Controller {
 				'X-Scope' => 'with_roles'
 			]
 		]);
-
 		$profileData = json_decode($profileResponse->getBody());
 
-		$mappedGroupIDs = $this->provisioningService->getMappedGroups($profileData);
-		if (in_array('block_unmapped', $generalSettings['options']) && count($mappedGroupIDs) === 0) {
+		$eventParticipationsData = null;
+		if ($this->settingsService->hasGeneralOption(SettingsService::GENERAL_OPTION_ENABLE_EVENT_MAPPING)) {
+			$eventParticipationsResponse = $client->get("{$baseUrl}/api/event_participations?filter[participant_id]={$profileData->id}&filter[participant_type]=Person&include=roles&fields[event_participations]=event_id,roles&fields[event_roles]=type",
+				[
+					'headers' => [
+						'Accept' => 'application/json',
+						'Authorization' => "Bearer $accessToken"
+					]
+				]);
+			$eventParticipationsData = json_decode($eventParticipationsResponse->getBody());
+		}
+
+		$mappedGroupIDs = $this->provisioningService->getMappedGroups($profileData, $eventParticipationsData);
+		if ($this->settingsService->hasGeneralOption(SettingsService::GENERAL_OPTION_BLOCK_UNMAPPED) && count($mappedGroupIDs) === 0) {
 			$this->logger->warning(
 				'User has no mapped groups and block users without groups is enabled',
 				['hitobito_id' => $profileData->id]
@@ -221,7 +234,7 @@ class LoginController extends Controller {
 		}
 		if (isset($existingUsers[0])) {
 			$user = $this->userManager->get($existingUsers[0]);
-		} elseif (in_array('email_lookup', $generalSettings['options'])) {
+		} elseif ($this->settingsService->hasGeneralOption(SettingsService::GENERAL_OPTION_EMAIL_LOOKUP)) {
 			$usersWithEmail = $this->userManager->getByEmail($profileData->email);
 			if (count($usersWithEmail) > 1) {
 				$this->logger->error('Multiple users found for email', ['email' => $profileData->email]);
@@ -237,7 +250,7 @@ class LoginController extends Controller {
 			}
 		}
 
-		$user = $this->provisioningService->provisionUser($profileData, $mappedGroupIDs, $user);
+		$user = $this->provisioningService->provisionUser($profileData, $user);
 
 		if (!$this->authenticateUser($user)) {
 			return $this->buildErrorTemplateResponse(